+++ R/server.R
@@ -31,24 +31,46 @@ registerClient <- function(client) {
 
 #' Resource Publishing
 #'
-#' Adds a directory of static resources to Shiny's web server, with the given
-#' path prefix. Primarily intended for package authors to make supporting
-#' JavaScript/CSS files available to their components.
+#' Add, remove, or list directory of static resources to Shiny's web server,
+#' with the given path prefix. Primarily intended for package authors to make
+#' supporting JavaScript/CSS files available to their components.
+#'
+#' Shiny provides two ways of serving static files (i.e., resources):
+#'
+#' 1. Static files under the `www/` directory are automatically made available
+#' under a request path that begins with `/`.
+#' 2. `addResourcePath()` makes static files in a `directoryPath` available
+#' under a request path that begins with `prefix`.
+#'
+#' The second approach is primarily intended for package authors to make
+#' supporting JavaScript/CSS files available to their components.
+#'
+#' Tools for managing static resources published by Shiny's web server:
+#'  * `addResourcePath()` adds a directory of static resources.
+#'  * `resourcePaths()` lists the currently active resource mappings.
+#'  * `removeResourcePath()` removes a directory of static resources.
 #'
 #' @param prefix The URL prefix (without slashes). Valid characters are a-z,
-#'   A-Z, 0-9, hyphen, period, and underscore.
-#'   For example, a value of 'foo' means that any request paths that begin with
-#'   '/foo' will be mapped to the given directory.
+#'   A-Z, 0-9, hyphen, period, and underscore. For example, a value of 'foo'
+#'   means that any request paths that begin with '/foo' will be mapped to the
+#'   given directory.
 #' @param directoryPath The directory that contains the static resources to be
 #'   served.
 #'
+#' @rdname resourcePaths
 #' @seealso [singleton()]
 #'
 #' @examples
 #' addResourcePath('datasets', system.file('data', package='datasets'))
+#' resourcePaths()
+#' removeResourcePath('datasets')
+#' resourcePaths()
+#'
+#' # make sure all resources are removed
+#' lapply(names(resourcePaths()), removeResourcePath)
 #' @export
 addResourcePath <- function(prefix, directoryPath) {
-  prefix <- prefix[1]
+  if (length(prefix) != 1) stop("prefix must be of length 1")
   if (!grepl('^[a-z0-9\\-_][a-z0-9\\-_.]*$', prefix, ignore.case = TRUE, perl = TRUE)) {
     stop("addResourcePath called with invalid prefix; please see documentation")
   }
@@ -62,6 +84,26 @@ addResourcePath <- function(prefix, directoryPath) {
         "`prefix` = '", prefix, "'; `directoryPath` = '" , directoryPath, "'")
     }
   )
+
+  # # Often times overwriting a resource path is "what you want",
+  # # but sometimes it can lead to difficult to diagnose issues
+  # # (e.g. an implict dependency might set a resource path that
+  # # conflicts with what you, the app author, are trying to register)
+  # # Note that previous versions of shiny used to warn about this case,
+  # # but it was eventually removed since it caused confusion (#567).
+  # # It seems a good compromise is to throw a more information message.
+  # if (getOption("shiny.resourcePathChanges", FALSE) &&
+  #     prefix %in% names(.globals$resourcePaths)) {
+  #   existingPath <- .globals$resourcePaths[[prefix]]$path
+  #   if (normalizedPath != existingPath) {
+  #     message(
+  #       "The resource path '", prefix, "' used to point to ",
+  #       existingPath, ", but it now points to ", normalizedPath, ". ",
+  #       "If your app doesn't work as expected, you may want to ",
+  #       "choose a different prefix name."
+  #     )
+  #   }
+  # }
 
   # If a shiny app is currently running, dynamically register this path with
   # the corresponding httpuv server object.
@@ -81,6 +123,33 @@ addResourcePath <- function(prefix, directoryPath) {
     func = staticHandler(normalizedPath)
   )
 }
+
+#' @rdname resourcePaths
+#' @export
+resourcePaths <- function() {
+  urls <- names(.globals$resourcePaths)
+  paths <- vapply(.globals$resourcePaths, function(x) x$path, character(1))
+  stats::setNames(paths, urls)
+}
+
+hasResourcePath <- function(prefix) {
+  prefix %in% names(resourcePaths())
+}
+
+#' @rdname resourcePaths
+#' @export
+removeResourcePath <- function(prefix) {
+  if (length(prefix) > 1) stop("`prefix` must be of length 1.")
+  if (!hasResourcePath(prefix)) {
+    warning("Resource ", prefix, " not found.")
+    return(invisible(FALSE))
+  }
+  .globals$resourcePaths[[prefix]] <- NULL
+  .globals$resources[[prefix]] <- NULL
+  invisible(TRUE)
+}
+
+
 
 # This function handles any GET request with two or more path elements where the
 # first path element matches a prefix that was previously added using
@@ -459,6 +528,49 @@ startApp <- function(appObj, port, host, quiet) {
     ),
     .globals$resourcePaths
   )
+
+  # throw an informative warning if a subdirectory of the
+  # app's www dir conflicts with another resource prefix
+  wwwDir <- httpuvApp$staticPaths[["/"]]$path
+  if (length(wwwDir)) {
+    # although httpuv allows for resource prefixes like 'foo/bar',
+    # we won't worry about conflicts in sub-sub directories since
+    # addResourcePath() currently doesn't allow it
+    wwwSubDirs <- list.dirs(wwwDir, recursive = FALSE, full.names = FALSE)
+    resourceConflicts <- intersect(wwwSubDirs, names(httpuvApp$staticPaths))
+    if (length(resourceConflicts)) {
+      warning(
+        "Found subdirectories of your app's www/ directory that ",
+        "conflict with other resource URL prefixes. ",
+        "Consider renaming these directories: '",
+        paste0("www/", resourceConflicts, collapse = "', '"), "'",
+        call. = FALSE
+      )
+    }
+  }
+
+  # check for conflicts in each pairwise combinations of resource mappings
+  checkResourceConflict <- function(paths) {
+    if (length(paths) < 2) return(NULL)
+    # ensure paths is a named character vector: c(resource_path = local_path)
+    paths <- vapply(paths, function(x) if (inherits(x, "staticPath")) x$path else x, character(1))
+    # get all possible pairwise combinations of paths
+    pair_indices <- utils::combn(length(paths), 2, simplify = FALSE)
+    lapply(pair_indices, function(x) {
+      p1 <- paths[x[1]]
+      p2 <- paths[x[2]]
+      if (identical(names(p1), names(p2)) && (p1 != p2)) {
+        warning(
+          "Found multiple local file paths pointing the same resource prefix: ", names(p1), ". ",
+          "If you run into resource-related issues (e.g. 404 requests), consider ",
+          "using `addResourcePath()` and/or `removeResourcePath()` to manage resource mappings.",
+          call. = FALSE
+        )
+      }
+    })
+  }
+  checkResourceConflict(httpuvApp$staticPaths)
+
   httpuvApp$staticPathOptions <- httpuv::staticPathOptions(
     html_charset = "utf-8",
     headers = list("X-UA-Compatible" = "IE=edge,chrome=1"),
@@ -875,7 +987,6 @@ runApp <- function(appDir=getwd(),
     captureStackTraces({
       while (!.globals$stopped) {
         ..stacktracefloor..(serviceApp())
-        Sys.sleep(0.001)
       }
     })
   )