+++ R/utils.R
@@ -173,8 +173,8 @@ anyUnnamed <- function(x) {
 }
 
 
-# Given a vector/list, returns a named vector (the labels will be blank).
-asNamedVector <- function(x) {
+# Given a vector/list, returns a named vector/list (the labels will be blank).
+asNamed <- function(x) {
   if (is.null(names(x))) {
     names(x) <- character(length(x))
   }
@@ -1800,3 +1800,8 @@ constantTimeEquals <- function(raw1, raw2) {
 
   sum(as.integer(xor(raw1, raw2))) == 0
 }
+
+cat_line <- function(...) {
+  cat(paste(..., "\n", collapse = ""))
+}
+