+++ inst/www/shared/shiny.js
@@ -1600,13 +1600,15 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
       function getTabIndex($tabset, tabsetId) {
         // The 0 is to ensure this works for empty tabsetPanels as well
         var existingTabIds = [0];
-        var leadingHref = "#tab-" + tabsetId + "-";
         // loop through all existing tabs, find the one with highest id
         // (since this is based on a numeric counter), and increment
         $tabset.find("> li").each(function () {
           var $tab = $(this).find("> a[data-toggle='tab']");
           if ($tab.length > 0) {
-            var index = $tab.attr("href").replace(leadingHref, "");
+            // remove leading url if it exists. (copy of bootstrap url stripper)
+            var href = $tab.attr("href").replace(/.*(?=#[^\s]+$)/, '');
+            // remove tab id to get the index
+            var index = href.replace("#tab-" + tabsetId + "-", "");
             existingTabIds.push(Number(index));
           }
         });
@@ -4918,18 +4920,33 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
       if (date === undefined) return;
       if (date === null) {
         $(el).bsDatepicker('setStartDate', null);
-      } else {
-        date = this._newDate(date);
-        date = this._UTCDateAsLocal(date);
-        if (!isNaN(date)) {
-          // Workaround for https://github.com/eternicode/bootstrap-datepicker/issues/2010
-          // If the start date when there's a two-digit year format, it will set
-          // the date value to null. So we'll save the value, set the start
-          // date, and the restore the value.
-          var curValue = $(el).bsDatepicker('getUTCDate');
-          $(el).bsDatepicker('setStartDate', date);
-          $(el).bsDatepicker('setUTCDate', curValue);
-        }
+        return;
+      }
+
+      date = this._newDate(date);
+      // If date parsing fails, do nothing
+      if (date === null) return;
+
+      date = this._UTCDateAsLocal(date);
+      if (isNaN(date)) return;
+      // Workaround for https://github.com/eternicode/bootstrap-datepicker/issues/2010
+      // If the start date when there's a two-digit year format, it will set
+      // the date value to null. So we'll save the value, set the start
+      // date, and the restore the value.
+      var curValue = $(el).bsDatepicker('getUTCDate');
+      $(el).bsDatepicker('setStartDate', date);
+      $(el).bsDatepicker('setUTCDate', curValue);
+
+      // Workaround for https://github.com/rstudio/shiny/issues/2335
+      // We only set the start date *after* the value in this special
+      // case so we don't effect the intended behavior of having a blank
+      // value when it falls outside the start date
+      if (typeof date.toDateString !== 'function') return;
+      if (typeof curValue.toDateString !== 'function') return;
+      if (date.toDateString() === curValue.toDateString()) {
+        $(el).bsDatepicker('setStartDate', null);
+        $(el).bsDatepicker('setUTCDate', curValue);
+        $(el).bsDatepicker('setStartDate', date);
       }
     },
     // Given an unambiguous date string or a Date object, set the max (end) date
@@ -4938,15 +4955,28 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
       if (date === undefined) return;
       if (date === null) {
         $(el).bsDatepicker('setEndDate', null);
-      } else {
-        date = this._newDate(date);
-        date = this._UTCDateAsLocal(date);
-        if (!isNaN(date)) {
-          // Workaround for same issue as in _setMin.
-          var curValue = $(el).bsDatepicker('getUTCDate');
-          $(el).bsDatepicker('setEndDate', date);
-          $(el).bsDatepicker('setUTCDate', curValue);
-        }
+        return;
+      }
+
+      date = this._newDate(date);
+      // If date parsing fails, do nothing
+      if (date === null) return;
+
+      date = this._UTCDateAsLocal(date);
+      if (isNaN(date)) return;
+
+      // Workaround for same issue as in _setMin.
+      var curValue = $(el).bsDatepicker('getUTCDate');
+      $(el).bsDatepicker('setEndDate', date);
+      $(el).bsDatepicker('setUTCDate', curValue);
+
+      // Workaround for same issue as in _setMin.
+      if (typeof date.toDateString !== 'function') return;
+      if (typeof curValue.toDateString !== 'function') return;
+      if (date.toDateString() === curValue.toDateString()) {
+        $(el).bsDatepicker('setEndDate', null);
+        $(el).bsDatepicker('setUTCDate', curValue);
+        $(el).bsDatepicker('setEndDate', date);
       }
     },
     // Given a date string of format yyyy-mm-dd, return a Date object with