+++ tests/test-helpers/app3-badglobal/global.R
@@ -1,0 +1,2 @@
+
+stop("I wasn't supposed to be sourced")