+++ tests/test-helpers/app4-both/r/lower.R
@@ -1,0 +1,1 @@
+lowerHelper <- 123