+++ tests/test-helpers/app5-bad-supporting/global.R
@@ -1,0 +1,2 @@
+
+happy <- TRUE